Move ConfigHelper to its own file
